refactor: move errors handlers in the proper file
