+++ renku/service/views/graph.py
@@ -21,8 +21,8 @@ from flask import request
 from renku.service.config import SERVICE_PREFIX
 from renku.service.controllers.graph_export import GraphExportCtrl
 from renku.service.views.api_versions import V1_0, V1_1, VersionedBlueprint
-from renku.service.views.decorators import accepts_json, handle_common_except, optional_identity, requires_cache
-from renku.service.views.error_handlers import handle_graph_errors
+from renku.service.views.decorators import accepts_json, optional_identity, requires_cache
+from renku.service.views.error_handlers import handle_common_except, handle_graph_errors
 
 GRAPH_BLUEPRINT_TAG = "graph"
 graph_blueprint = VersionedBlueprint(GRAPH_BLUEPRINT_TAG, __name__, url_prefix=SERVICE_PREFIX)
